Update module name to CustomPhotoGenerator, refactored the code to contain server, utils, telegramclient, model modules.
