+++ utils/getImage.go
@@ -1,0 +1,19 @@
+package utils
+
+import (
+	"image"
+	"net/http"
+)
+
+func GetImage(url string) (image.Image, error) {
+	res, err := http.Get(url)
+	if err != nil {
+		return nil, err
+	}
+	defer res.Body.Close()
+	img, _, err := image.Decode(res.Body)
+	if err != nil {
+		return nil, err
+	}
+	return img, nil
+}
